Renamed all js files to typescript
