+++ components/WaveBlueStar.tsx
