+++ components/WavePinkStar.tsx
