+++ pages/index.tsx
@@ -1,8 +1,9 @@
 import Head from "next/head"
 
-import DesktopNavbar from "../components/DesktopNavbar"
 import Container from "../components/Container"
 import HeroSection from "../components/HeroSection"
+import DesktopNavbar from "../components/DesktopNavbar"
+import ProjectSection from "../components/ProjectSection"
 
 export default function Home() {
   return (
@@ -15,10 +16,10 @@ export default function Home() {
         />
         <link rel="icon" href="/Logo.svg" />
       </Head>
-
       <DesktopNavbar />
       <Container>
         <HeroSection></HeroSection>
+        <ProjectSection />
       </Container>
     </div>
   )
